+++ src/main/kotlin/lean4ij/project/listeners/LeanFileCaretListener.kt
@@ -9,6 +9,7 @@ import com.intellij.openapi.editor.event.CaretListener
 import com.intellij.openapi.editor.ex.EditorEx
 import com.intellij.openapi.project.Project
 import com.intellij.util.ui.JBUI
+import lean4ij.util.LeanUtil
 import java.beans.PropertyChangeEvent
 import java.beans.PropertyChangeListener
 
@@ -22,6 +23,9 @@ class LeanFileCaretListener(private val project: Project) : CaretListener/*, Pro
      */
     override fun caretPositionChanged(event: CaretEvent) {
         // TODO passing things like editor etc seems cumbersome, maybe add some implement for context
+        if (!LeanUtil.isLeanFile(event.editor.virtualFile.path)) {
+            return
+        }
         leanProjectService.file(event.editor.virtualFile).updateCaret(event.editor, event.newPosition)
     }