+++ src/main/kotlin/lean4ij/util/LeanUtil.kt
@@ -1,5 +1,6 @@
 package lean4ij.util
 
+import com.intellij.openapi.vfs.VirtualFile
 import java.io.File
 import java.io.IOException
 import java.util.concurrent.TimeUnit
@@ -35,8 +36,12 @@ object LeanUtil {
         }
     }
 
+    fun isLeanFile(file: VirtualFile) : Boolean {
+        return file.extension?.let { it == "lean" || it == "lean4"} ?: false
+    }
+
     fun isLeanFile(url: String) : Boolean {
-        return url.endsWith("lean") || url.endsWith("lean4")
+        return url.endsWith(".lean") || url.endsWith(".lean4")
     }
 
 }